+++ internal/model/wechat_reply.gen.go
@@ -14,7 +14,7 @@ type WechatReply struct {
 	Keywords  string                `gorm:"column:keywords;type:varchar(64);not null;comment:关键字" json:"keywords"`                          // 关键字
 	Type      string                `gorm:"column:type;type:varchar(32);not null;comment:回复类型" json:"type"`                                 // 回复类型
 	Data      string                `gorm:"column:data;type:text;not null;comment:回复数据" json:"data"`                                        // 回复数据
-	Status    int32                 `gorm:"column:status;type:tinyint unsigned;not null;default:1;comment:回复状态 0=不可用  1 =可用" json:"status"` // 回复状态 0=不可用  1 =可用
+	Status    int64                 `gorm:"column:status;type:tinyint unsigned;not null;default:1;comment:回复状态 0=不可用  1 =可用" json:"status"` // 回复状态 0=不可用  1 =可用
 	CreatedAt int64                 `gorm:"column:created_at;type:bigint" json:"created_at"`
 	UpdatedAt int64                 `gorm:"column:updated_at;type:bigint" json:"updated_at"`
 	DeletedAt soft_delete.DeletedAt `gorm:"column:deleted_at;type:bigint" json:"-"`
